test(skills): split remote reconciliation coverage
